feat: Add base & collection_list template
This commit will add the base.html template. It contains a draft
version for the base file. This will be modified later, once
the functionality is granted.

An empty collection_list.html template is added, since our
CollectionListView is referencing to it. Will be styled later.

Some comments have been added to various files.