feat: Update code to display list of collections
This commit will update the HomeView and CollectionListView in order
that if a user is authenticated, he will be redirected to the list of
collections.
Redirect url in settings file is also updated.
The collection_list.html template got enhanced in order to show the
functionality, however, still very basic. Will be updated later.