feat: Add functionality to create, list collection
This commit will add a new CollectionCreateView in order to create
a new collection. It will reference to the collection_list once a
new collection is created. All collections are then be visible.

It will add a new collection_create.html template, which is very
basic and will be updated later.

Most urls and references are updated. The old and unnecessary
index.html site, which was copied for reference, is deleted.