feat: Add feature to have default collection image
This commit will add the feature to have a default image in the
collection list, in case a user does not upload his own image.

It extens the template with this functionality and also includes
necessary changes on the settings.py file in order to process
the "static" template tag.