feat: Add collection detail view
This commit will add the functionality of clicking on the view button
on one of the collections in the collection list, and you get
redirected to the list of items in that particular collection.
If there are no items, a text will tell you and you can create one.

Styling is still very raw. This commit updates all related components
for this functionality, including adding new template.