feat: Add function to update & delete collection
This commit will add two more views.
- Update view, in order to update an existing collection for name,
description or the image / replace the default image.

- Delete view, in order to delete a full collection, including a
security question to be sure for deletion. Since CASCADE is used in
the model, it will delete all items.

Templates, very generic, are also added as well as the existing
collection_list template got enhanced with the two more buttons.