refactor: Simplify item model
This commit will refactor the item model. I had to step back from the
idea of having multiple images per item, since this has increased the
level of complexity for the whole project quite a lot.
Since I have to deliver MVP, I will let this feature for future
implementation. I had to adapt the db model to reflect this.

This commit will remove the image model and add the image field to
the item model. Therefore a user will be able to upload one image
per item.