feat: Add create item template
This commit will add the create item template to look like the create
collection template. It is still basic and raw, but it works and
has the same look.
It is now possible to create a new item, upload an image or get a
default image and view the items in collection_detail view.