feat: Add functional navigatio buttons
This commit will add a few functional navigation buttons.
In the collection detail view, it will add buttons to edit the
collection, create an item and go back to my collections

In the collection create view, it will add a cancel button to go back
to the collection list.

In the item create view, it will add a cancel button to go back to
the collection list, where the user came from.