feat: Add search item functionality
This commit will add a search functionality inside the collection_detail
view. It let's the user search for the given fields inside a
particular collection and only in that collection.

If search is successful, it wil display the result, if nothing is
found, it will display a text and a cancel button to return to the
collection where the user came from.

The feature is implemented inside / to work with pagination.