feat: Add contact app - not functional yet
This commit will add the new contact app including the necessary
folder structure. It will also include the settings of the main
app urls.py as well as the settings.py

The contact app contains a new forms.py file, with a simple form.
The views.py got created and there are two Classes:
- The SendEmailView to wire up the form and create the data to be sent
- The EmailSuccessView to point to the success url.

Furthermore, it contains an urls.py file with the wiring of the two
classes to the contact.html and the success.html file.
The two .html file contan the very basic form data and success message.

Note that with this commit, the contact form is not functional, the
SendGrid Email Provider first needs to be wired up.