feat: Add Mailtrap as contact mail provider
This commit will add all settings for Mailtrap as testemail provider.
The settings.py got updated as well as the env.py, since the
credentials should not be exposed.

The contact.html and the success.html. got updated to be responsive,
but still very raw, needs to be updated later, when the whole app
will be styled.

The widget attribute has been used in the ContactForm class, in order
to set the placeholder in the form. Therefore the class ContactForm
in my custom forms.py file got updated.

Whole service is validated and working.