feat: Add About site and wire it up
This commit will add a simple "about" site to the app. A very simple
view has been created inside the collection app, since it is not worth
to create a new app just for this simple site.
The urls.py and the navbar got updated too.

Like everything else, this site is only available to logged in users.