refactor: Modify error handling & file structure
This commit will change the CBV for the error handling into
FBV. It will also move the error sites to a new folder in the root
app, called folders.
Since error sites need to be available for all apps, I had to move
them up in the hierarchy. Furthermore, I could not bring it up and
running with the CBV whne having in the root, so I decided to go
for FBV.
The CBV for the errors got deleted and in a new views.py of the main
app, the FBV are crated.
urls of the main app got enhanced with the error handler.