refactor: Modify template structure of collection
This commit will modify the template structure of the collection app.
It will move the templates for the collection app to a new folder
templates/collection inside the collection app. This is the proper
whay to store the files.
views have been updated with the new path for the templates