refactor: Modify structure of contact app
This commit will modify the template structure of the contact app.
It will move the templates for the contact app to a new folder
templates/contact inside the contact app. This is the proper
whay to store the files.
views have been updated with the new path for the templates