feat: Add footer via includes option
This commit will add a footer, to some of the sites. It will add
it to the entry main site, once a user arrives and is unauthorised.
Furthermore it is added to the about site and the contact site.

The injection is done via the includes feature. Templates have been
updated due to this.

A custom css file has been added.
Styling still raw, will be finished later.