refactor: Prevent email in contact from exposing
This commit will refactor the existing contact form settings in the
way, that it will no longer expose the EMAIL_RECEIVE mail address
to github. I have added it as an env var, also on heroku.
The setting is used to see in mailtrapp, to where the mail should
be forwarded. It is embedded int the particular view.