style: Add style to collection detail & overall
This commit will style the collection detail site in order to be inline
with the overall look and feel.
It will also change / limit the size of the picuters for collection
and items, in order to have the cards the same height. Text
truncation for the collection / item description is also introduced,
for the same reason.
Background color of the card-text has been changed in order to have
contrast, if a image has white background.
The buttons to edit a collection have also been changed to small size,
furthermore the container-fluid attribute has been removed from the
overall content container...I did not like the look on certian screen
sizes.